pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 79
target-version = ['py38']

[tool.poetry]
name = "idis"
version = "0.1.0"
description = "image de-identification service that uses DICOM Confidentiality options"
authors = ["Sjoerd Kerkstra <w.s.kerkstra@pm.me>"]
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.7"
celery = "*"
redis = "*"
django = "<3.1"
django-countries = "*"
django-crispy-forms = "*"
django-userena-ce = "*"
djangorestframework = "*"
docker = "*"
"oauth2" = "*"
python-magic = "*"
python-memcached = "*"
pytz = "*"
social-auth-app-django = "*"
gunicorn = "*"
uvicorn = "*"
django-celery-email = "*"
django-favicon-plus = "*"
"psycopg2" = "*"
"django-select2" = "*"
django-celery-results = "*"
django-autocomplete-light = "*"
django-storages = "*"
whitenoise = "*"
djangorestframework-guardian = "*"
django-extensions = "*"
django-simple-history = "*"
sentry-sdk = "*"
django-cors-headers = "*"
django-speedinfo = "*"
drf-yasg = {version = "*", extras = ["validation"]}
pydicom = "*"
django-celery-beat = "!=1.6.0"
django-guardian = "!=2.2.0"
django-markdownx = "^3.0.1"
django-filter = "^2.2.0"
django-allauth = "*"
idissend = "^0.2.13"
anonapi = "^1.1.2"

[tool.poetry.dev-dependencies]
pytest-django = "*"
pytest-cov = "*"
pytest-mock = "*"
factory-boy = "*"
django-debug-toolbar = "*"
sphinx-autobuild = "*"
sphinx = "*"
pyupgrade = "*"
pytest-xdist = "*"
sphinx-autodoc-typehints = "*"
werkzeug = "*"
sphinx-rtd-theme = "*"
